+++ src/core/models/tablero/Test_Tablero.py
@@ -9,7 +9,7 @@ class TestTablero(unittest.TestCase):
         self.tablero_vacio = [[] for _ in range(24)]
         self.tablero = Tablero(self.tablero_vacio)
 
-    def test_mover_ficha_a_espacio_vacio(self):
+    def test_mover_ficha_a_espacio_vacio(self): #mover_ficha()
         ficha = Ficha(TipoFicha.NEGRA.value)
         self.tablero.__tablero__[0].append(ficha)
         
@@ -41,6 +41,20 @@ class TestTablero(unittest.TestCase):
         
         with self.assertRaises(CasillaOcupadaException):
             self.tablero.mover_ficha(ficha_negra, 0, 1)
-
+    def test_mover_ficha_a_espacio_con_una_ficha_rival(self):
+        ficha_negra = Ficha(TipoFicha.NEGRA.value)
+        ficha_roja = Ficha(TipoFicha.ROJA.value)
+        
+        self.tablero.__tablero__[0].append(ficha_negra)
+        self.tablero.__tablero__[1].append(ficha_roja)
+        
+        self.tablero.mover_ficha(ficha_negra, 0, 1)
+        
+        self.assertEqual(len(self.tablero.__tablero__[0]), 0)
+        self.assertEqual(len(self.tablero.__tablero__[1]), 1)
+        self.assertEqual(self.tablero.__tablero__[1][0], ficha_negra)
+        self.assertEqual(len(self.tablero.fichas_comidas), 1) #ficha roja
+        self.assertEqual(self.tablero.fichas_comidas[0], ficha_roja)
+        self.assertTrue(self.tablero.fichas_comidas[0].comida)
 if __name__ == '__main__':
     unittest.main()