+++ CHANGELOG.md
@@ -53,3 +53,16 @@ Clase tablero, clase tablero validador, testeadas
 ### Arreglado
 
 ### Eliminado
+
+## [1.0.5] - 24-08-2025
+
+### Agregado
+-metodos para saber si come o no una ficha en moviemiento agregadas
+-setter de comida en ficha agregado
+-guardar fichas en "medio" del tablero cuando se comen
+-tests
+-comentarios agregados en test
+### Cambiado
+### Arreglado
+
+### Eliminado